Added GetDecimalFactions call and related public methods and tests

+++ helloserve.com.RandomOrg/Models/RandomData.cs
@@ -8,7 +8,7 @@ namespace helloserve.com.RandomOrg.Models
 {
     internal class RandomData
     {
-        public decimal[] data { get; set; }
+        public double[] data { get; set; }
         public string completionTime { get; set; }
     }
 }
